Restructure StartPage
